+++ tests/index.test.ts
@@ -15,6 +15,8 @@ import {
   ResultAsync,
 } from '../src'
 
+import { vitest, describe, expect, it } from 'vitest'
+
 describe('Result.Ok', () => {
   it('Creates an Ok value', () => {
     const okVal = ok(12)
@@ -47,7 +49,7 @@ describe('Result.Ok', () => {
 
   it('Maps over an Ok value', () => {
     const okVal = ok(12)
-    const mapFn = jest.fn((number) => number.toString())
+    const mapFn = vitest.fn((number) => number.toString())
 
     const mapped = okVal.map(mapFn)
 
@@ -57,7 +59,7 @@ describe('Result.Ok', () => {
   })
 
   it('Skips `mapErr`', () => {
-    const mapErrorFunc = jest.fn((_error) => 'mapped error value')
+    const mapErrorFunc = vitest.fn((_error) => 'mapped error value')
 
     const notMapped = ok(12).mapErr(mapErrorFunc)
 
@@ -92,7 +94,7 @@ describe('Result.Ok', () => {
 
       expect(flattened.isOk()).toBe(false)
 
-      const nextFn = jest.fn((_val) => ok('noop'))
+      const nextFn = vitest.fn((_val) => ok('noop'))
 
       flattened.andThen(nextFn)
 
@@ -103,7 +105,7 @@ describe('Result.Ok', () => {
   describe('andThrough', () => {
     it('Calls the passed function but returns an original ok', () => {
       const okVal = ok(12)
-      const passedFn = jest.fn((_number) => ok(undefined))
+      const passedFn = vitest.fn((_number) => ok(undefined))
 
       const thrued = okVal.andThrough(passedFn)
       expect(thrued.isOk()).toBe(true)
@@ -124,7 +126,7 @@ describe('Result.Ok', () => {
       expect(thrued.isOk()).toBe(false)
       expect(thrued._unsafeUnwrapErr()).toStrictEqual('Whoopsies!')
 
-      const nextFn = jest.fn((_val) => ok('noop'))
+      const nextFn = vitest.fn((_val) => ok('noop'))
 
       thrued.andThen(nextFn)
 
@@ -135,7 +137,7 @@ describe('Result.Ok', () => {
   describe('andTee', () => {
     it('Calls the passed function but returns an original ok', () => {
       const okVal = ok(12)
-      const passedFn = jest.fn((_number) => {})
+      const passedFn = vitest.fn((_number) => {})
 
       const teed = okVal.andTee(passedFn)
 
@@ -145,7 +147,9 @@ describe('Result.Ok', () => {
     })
     it('returns an original ok even when the passed function fails', () => {
       const okVal = ok(12)
-      const passedFn = jest.fn((_number) => { throw new Error('OMG!') })
+      const passedFn = vitest.fn((_number) => {
+        throw new Error('OMG!')
+      })
 
       const teed = okVal.andTee(passedFn)
 
@@ -155,10 +159,35 @@ describe('Result.Ok', () => {
     })
   })
 
+  describe('orTee', () => {
+    it('Calls the passed function but returns an original err', () => {
+      const errVal = err(12)
+      const passedFn = vitest.fn((_number) => {})
+
+      const teed = errVal.orTee(passedFn)
+
+      expect(teed.isErr()).toBe(true)
+      expect(passedFn).toHaveBeenCalledTimes(1)
+      expect(teed._unsafeUnwrapErr()).toStrictEqual(12)
+    })
+    it('returns an original err even when the passed function fails', () => {
+      const errVal = err(12)
+      const passedFn = vitest.fn((_number) => {
+        throw new Error('OMG!')
+      })
+
+      const teed = errVal.orTee(passedFn)
+
+      expect(teed.isErr()).toBe(true)
+      expect(passedFn).toHaveBeenCalledTimes(1)
+      expect(teed._unsafeUnwrapErr()).toStrictEqual(12)
+    })
+  })
+
   describe('asyncAndThrough', () => {
     it('Calls the passed function but returns an original ok as Async', async () => {
       const okVal = ok(12)
-      const passedFn = jest.fn((_number) => okAsync(undefined))
+      const passedFn = vitest.fn((_number) => okAsync(undefined))
 
       const teedAsync = okVal.asyncAndThrough(passedFn)
       expect(teedAsync).toBeInstanceOf(ResultAsync)
@@ -182,7 +211,7 @@ describe('Result.Ok', () => {
       expect(teed.isOk()).toBe(false)
       expect(teed._unsafeUnwrapErr()).toStrictEqual('Whoopsies!')
 
-      const nextFn = jest.fn((_val) => ok('noop'))
+      const nextFn = vitest.fn((_val) => ok('noop'))
 
       teed.andThen(nextFn)
 
@@ -192,7 +221,7 @@ describe('Result.Ok', () => {
   describe('orElse', () => {
     it('Skips orElse on an Ok value', () => {
       const okVal = ok(12)
-      const errorCallback = jest.fn((_errVal) => err<number, string>('It is now a string'))
+      const errorCallback = vitest.fn((_errVal) => err<number, string>('It is now a string'))
 
       expect(okVal.orElse(errorCallback)).toEqual(ok(12))
       expect(errorCallback).not.toHaveBeenCalled()
@@ -223,7 +252,7 @@ describe('Result.Ok', () => {
   })
 
   it('Maps to a promise', async () => {
-    const asyncMapper = jest.fn((_val) => {
+    const asyncMapper = vitest.fn((_val) => {
       // ...
       // complex logic
       // ..
@@ -249,8 +278,8 @@ describe('Result.Ok', () => {
   })
 
   it('Matches on an Ok', () => {
-    const okMapper = jest.fn((_val) => 'weeeeee')
-    const errMapper = jest.fn((_val) => 'wooooo')
+    const okMapper = vitest.fn((_val) => 'weeeeee')
+    const errMapper = vitest.fn((_val) => 'wooooo')
 
     const matched = ok(12).match(okMapper, errMapper)
 
@@ -294,7 +323,7 @@ describe('Result.Err', () => {
   it('Skips `map`', () => {
     const errVal = err('I am your father')
 
-    const mapper = jest.fn((_value) => 'noooo')
+    const mapper = vitest.fn((_value) => 'noooo')
 
     const hopefullyNotMapped = errVal.map(mapper)
 
@@ -306,7 +335,7 @@ describe('Result.Err', () => {
   it('Maps over an Err', () => {
     const errVal = err('Round 1, Fight!')
 
-    const mapper = jest.fn((error: string) => error.replace('1', '2'))
+    const mapper = vitest.fn((error: string) => error.replace('1', '2'))
 
     const mapped = errVal.mapErr(mapper)
 
@@ -323,7 +352,7 @@ describe('Result.Err', () => {
   it('Skips over andThen', () => {
     const errVal = err('Yolo')
 
-    const mapper = jest.fn((_val) => ok<string, string>('yooyo'))
+    const mapper = vitest.fn((_val) => ok<string, string>('yooyo'))
 
     const hopefullyNotFlattened = errVal.andThen(mapper)
 
@@ -335,7 +364,7 @@ describe('Result.Err', () => {
   it('Skips over andThrough', () => {
     const errVal = err('Yolo')
 
-    const mapper = jest.fn((_val) => ok<void, string>(undefined))
+    const mapper = vitest.fn((_val) => ok<void, string>(undefined))
 
     const hopefullyNotFlattened = errVal.andThrough(mapper)
 
@@ -347,7 +376,7 @@ describe('Result.Err', () => {
   it('Skips over andTee', () => {
     const errVal = err('Yolo')
 
-    const mapper = jest.fn((_val) => {})
+    const mapper = vitest.fn((_val) => {})
 
     const hopefullyNotFlattened = errVal.andTee(mapper)
 
@@ -359,7 +388,7 @@ describe('Result.Err', () => {
   it('Skips over asyncAndThrough but returns ResultAsync instead', async () => {
     const errVal = err('Yolo')
 
-    const mapper = jest.fn((_val) => okAsync<string, unknown>('Async'))
+    const mapper = vitest.fn((_val) => okAsync<string, unknown>('Async'))
 
     const hopefullyNotFlattened = errVal.asyncAndThrough(mapper)
     expect(hopefullyNotFlattened).toBeInstanceOf(ResultAsync)
@@ -373,7 +402,7 @@ describe('Result.Err', () => {
   it('Transforms error into ResultAsync within `asyncAndThen`', async () => {
     const errVal = err('Yolo')
 
-    const asyncMapper = jest.fn((_val) => okAsync<string, string>('yooyo'))
+    const asyncMapper = vitest.fn((_val) => okAsync<string, string>('yooyo'))
 
     const hopefullyNotFlattened = errVal.asyncAndThen(asyncMapper)
 
@@ -385,7 +414,7 @@ describe('Result.Err', () => {
   })
 
   it('Does not invoke callback within `asyncMap`', async () => {
-    const asyncMapper = jest.fn((_val) => {
+    const asyncMapper = vitest.fn((_val) => {
       // ...
       // complex logic
       // ..
@@ -411,8 +440,8 @@ describe('Result.Err', () => {
   })
 
   it('Matches on an Err', () => {
-    const okMapper = jest.fn((_val) => 'weeeeee')
-    const errMapper = jest.fn((_val) => 'wooooo')
+    const okMapper = vitest.fn((_val) => 'weeeeee')
+    const errMapper = vitest.fn((_val) => 'wooooo')
 
     const matched = err(12).match(okMapper, errMapper)
 
@@ -438,7 +467,7 @@ describe('Result.Err', () => {
   describe('orElse', () => {
     it('invokes the orElse callback on an Err value', () => {
       const okVal = err('BOOOM!')
-      const errorCallback = jest.fn((_errVal) => err(true))
+      const errorCallback = vitest.fn((_errVal) => err(true))
 
       expect(okVal.orElse(errorCallback)).toEqual(err(true))
       expect(errorCallback).toHaveBeenCalledTimes(1)
@@ -460,15 +489,15 @@ describe('Result.fromThrowable', () => {
 
   // Added for issue #300 -- the test here is not so much that expectations are met as that the test compiles.
   it('Accepts an inner function which takes arguments', () => {
-    const hello = (fname: string): string => `hello, ${fname}`;
-    const safeHello = Result.fromThrowable(hello);
+    const hello = (fname: string): string => `hello, ${fname}`
+    const safeHello = Result.fromThrowable(hello)
 
-    const result = hello('Dikembe');
-    const safeResult = safeHello('Dikembe');
+    const result = hello('Dikembe')
+    const safeResult = safeHello('Dikembe')
 
-    expect(safeResult).toBeInstanceOf(Ok);
-    expect(result).toEqual(safeResult._unsafeUnwrap());
-  });
+    expect(safeResult).toBeInstanceOf(Ok)
+    expect(result).toEqual(safeResult._unsafeUnwrap())
+  })
 
   it('Creates a function that returns an err when the inner function throws', () => {
     const thrower = (): string => {
@@ -503,7 +532,7 @@ describe('Result.fromThrowable', () => {
   })
 
   it('has a top level export', () => {
-      expect(fromThrowable).toBe(Result.fromThrowable)
+    expect(fromThrowable).toBe(Result.fromThrowable)
   })
 })
 
@@ -534,15 +563,15 @@ describe('Utils', () => {
       })
 
       it('Combines heterogeneous lists', () => {
-        type HeterogenousList = [ Result<string, string>, Result<number, number>, Result<boolean, boolean> ]
-
-        const heterogenousList: HeterogenousList = [
-          ok('Yooooo'),
-          ok(123),
-          ok(true),
+        type HeterogenousList = [
+          Result<string, string>,
+          Result<number, number>,
+          Result<boolean, boolean>,
         ]
 
-        type ExpecteResult = Result<[ string, number, boolean ], string | number | boolean>
+        const heterogenousList: HeterogenousList = [ok('Yooooo'), ok(123), ok(true)]
+
+        type ExpecteResult = Result<[string, number, boolean], string | number | boolean>
 
         const result: ExpecteResult = Result.combine(heterogenousList)
 
@@ -550,21 +579,18 @@ describe('Utils', () => {
       })
 
       it('Does not destructure / concatenate arrays', () => {
-        type HomogenousList = [
-          Result<string[], boolean>,
-          Result<number[], string>,
-        ]
+        type HomogenousList = [Result<string[], boolean>, Result<number[], string>]
 
-        const homogenousList: HomogenousList = [
-          ok(['hello', 'world']),
-          ok([1, 2, 3])
-        ]
+        const homogenousList: HomogenousList = [ok(['hello', 'world']), ok([1, 2, 3])]
 
-        type ExpectedResult = Result<[ string[], number[] ], boolean | string>
+        type ExpectedResult = Result<[string[], number[]], boolean | string>
 
         const result: ExpectedResult = Result.combine(homogenousList)
 
-        expect(result._unsafeUnwrap()).toEqual([ [ 'hello', 'world' ], [ 1, 2, 3 ]])
+        expect(result._unsafeUnwrap()).toEqual([
+          ['hello', 'world'],
+          [1, 2, 3],
+        ])
       })
     })
 
@@ -573,7 +599,7 @@ describe('Utils', () => {
         const asyncResultList = [okAsync(123), okAsync(456), okAsync(789)]
 
         const resultAsync: ResultAsync<number[], never[]> = ResultAsync.combine(asyncResultList)
-        
+
         expect(resultAsync).toBeInstanceOf(ResultAsync)
 
         const result = await ResultAsync.combine(asyncResultList)
@@ -608,14 +634,14 @@ describe('Utils', () => {
           okAsync('Yooooo'),
           okAsync(123),
           okAsync(true),
-          okAsync([ 1, 2, 3]),
+          okAsync([1, 2, 3]),
         ]
 
-        type ExpecteResult = Result<[ string, number, boolean, number[] ], string | number | boolean>
+        type ExpecteResult = Result<[string, number, boolean, number[]], string | number | boolean>
 
         const result: ExpecteResult = await ResultAsync.combine(heterogenousList)
 
-        expect(result._unsafeUnwrap()).toEqual(['Yooooo', 123, true, [ 1, 2, 3 ]])
+        expect(result._unsafeUnwrap()).toEqual(['Yooooo', 123, true, [1, 2, 3]])
       })
     })
   })
@@ -645,15 +671,15 @@ describe('Utils', () => {
       })
 
       it('Combines heterogeneous lists', () => {
-        type HeterogenousList = [ Result<string, string>, Result<number, number>, Result<boolean, boolean> ]
-
-        const heterogenousList: HeterogenousList = [
-          ok('Yooooo'),
-          ok(123),
-          ok(true),
+        type HeterogenousList = [
+          Result<string, string>,
+          Result<number, number>,
+          Result<boolean, boolean>,
         ]
 
-        type ExpecteResult = Result<[ string, number, boolean ], (string | number | boolean)[]>
+        const heterogenousList: HeterogenousList = [ok('Yooooo'), ok(123), ok(true)]
+
+        type ExpecteResult = Result<[string, number, boolean], (string | number | boolean)[]>
 
         const result: ExpecteResult = Result.combineWithAllErrors(heterogenousList)
 
@@ -661,21 +687,18 @@ describe('Utils', () => {
       })
 
       it('Does not destructure / concatenate arrays', () => {
-        type HomogenousList = [
-          Result<string[], boolean>,
-          Result<number[], string>,
-        ]
+        type HomogenousList = [Result<string[], boolean>, Result<number[], string>]
 
-        const homogenousList: HomogenousList = [
-          ok(['hello', 'world']),
-          ok([1, 2, 3])
-        ]
+        const homogenousList: HomogenousList = [ok(['hello', 'world']), ok([1, 2, 3])]
 
-        type ExpectedResult = Result<[ string[], number[] ], (boolean | string)[]>
+        type ExpectedResult = Result<[string[], number[]], (boolean | string)[]>
 
         const result: ExpectedResult = Result.combineWithAllErrors(homogenousList)
 
-        expect(result._unsafeUnwrap()).toEqual([ [ 'hello', 'world' ], [ 1, 2, 3 ]])
+        expect(result._unsafeUnwrap()).toEqual([
+          ['hello', 'world'],
+          [1, 2, 3],
+        ])
       })
     })
     describe('`ResultAsync.combineWithAllErrors`', () => {
@@ -703,15 +726,15 @@ describe('Utils', () => {
       })
 
       it('Combines heterogeneous lists', async () => {
-        type HeterogenousList = [ ResultAsync<string, string>, ResultAsync<number, number>, ResultAsync<boolean, boolean> ]
-
-        const heterogenousList: HeterogenousList = [
-          okAsync('Yooooo'),
-          okAsync(123),
-          okAsync(true),
+        type HeterogenousList = [
+          ResultAsync<string, string>,
+          ResultAsync<number, number>,
+          ResultAsync<boolean, boolean>,
         ]
 
-        type ExpecteResult = Result<[ string, number, boolean ], (string | number | boolean)[]>
+        const heterogenousList: HeterogenousList = [okAsync('Yooooo'), okAsync(123), okAsync(true)]
+
+        type ExpecteResult = Result<[string, number, boolean], (string | number | boolean)[]>
 
         const result: ExpecteResult = await ResultAsync.combineWithAllErrors(heterogenousList)
 
@@ -808,7 +831,7 @@ describe('ResultAsync', () => {
     it('Maps a value using a synchronous function', async () => {
       const asyncVal = okAsync(12)
 
-      const mapSyncFn = jest.fn((number) => number.toString())
+      const mapSyncFn = vitest.fn((number) => number.toString())
 
       const mapped = asyncVal.map(mapSyncFn)
 
@@ -824,7 +847,7 @@ describe('ResultAsync', () => {
     it('Maps a value using an asynchronous function', async () => {
       const asyncVal = okAsync(12)
 
-      const mapAsyncFn = jest.fn((number) => Promise.resolve(number.toString()))
+      const mapAsyncFn = vitest.fn((number) => Promise.resolve(number.toString()))
 
       const mapped = asyncVal.map(mapAsyncFn)
 
@@ -840,7 +863,7 @@ describe('ResultAsync', () => {
     it('Skips an error', async () => {
       const asyncErr = errAsync<number, string>('Wrong format')
 
-      const mapSyncFn = jest.fn((number) => number.toString())
+      const mapSyncFn = vitest.fn((number) => number.toString())
 
       const notMapped = asyncErr.map(mapSyncFn)
 
@@ -858,7 +881,7 @@ describe('ResultAsync', () => {
     it('Maps an error using a synchronous function', async () => {
       const asyncErr = errAsync('Wrong format')
 
-      const mapErrSyncFn = jest.fn((str) => 'Error: ' + str)
+      const mapErrSyncFn = vitest.fn((str) => 'Error: ' + str)
 
       const mappedErr = asyncErr.mapErr(mapErrSyncFn)
 
@@ -874,7 +897,7 @@ describe('ResultAsync', () => {
     it('Maps an error using an asynchronous function', async () => {
       const asyncErr = errAsync('Wrong format')
 
-      const mapErrAsyncFn = jest.fn((str) => Promise.resolve('Error: ' + str))
+      const mapErrAsyncFn = vitest.fn((str) => Promise.resolve('Error: ' + str))
 
       const mappedErr = asyncErr.mapErr(mapErrAsyncFn)
 
@@ -890,7 +913,7 @@ describe('ResultAsync', () => {
     it('Skips a value', async () => {
       const asyncVal = okAsync(12)
 
-      const mapErrSyncFn = jest.fn((str) => 'Error: ' + str)
+      const mapErrSyncFn = vitest.fn((str) => 'Error: ' + str)
 
       const notMapped = asyncVal.mapErr(mapErrSyncFn)
 
@@ -908,7 +931,7 @@ describe('ResultAsync', () => {
     it('Maps a value using a function returning a ResultAsync', async () => {
       const asyncVal = okAsync(12)
 
-      const andThenResultAsyncFn = jest.fn(() => okAsync('good'))
+      const andThenResultAsyncFn = vitest.fn(() => okAsync('good'))
 
       const mapped = asyncVal.andThen(andThenResultAsyncFn)
 
@@ -924,7 +947,7 @@ describe('ResultAsync', () => {
     it('Maps a value using a function returning a Result', async () => {
       const asyncVal = okAsync(12)
 
-      const andThenResultFn = jest.fn(() => ok('good'))
+      const andThenResultFn = vitest.fn(() => ok('good'))
 
       const mapped = asyncVal.andThen(andThenResultFn)
 
@@ -940,7 +963,7 @@ describe('ResultAsync', () => {
     it('Skips an Error', async () => {
       const asyncVal = errAsync<string, string>('Wrong format')
 
-      const andThenResultFn = jest.fn(() => ok<string, string>('good'))
+      const andThenResultFn = vitest.fn(() => ok<string, string>('good'))
 
       const notMapped = asyncVal.andThen(andThenResultFn)
 
@@ -963,7 +986,7 @@ describe('ResultAsync', () => {
         DB persistence (create or update)
         API calls (create or update)
       */
-      const andThroughResultAsyncFn = jest.fn(() => okAsync('good'))
+      const andThroughResultAsyncFn = vitest.fn(() => okAsync('good'))
 
       const thrued = asyncVal.andThrough(andThroughResultAsyncFn)
 
@@ -979,7 +1002,7 @@ describe('ResultAsync', () => {
     it('Maps to an error when map function returning ResultAsync fails', async () => {
       const asyncVal = okAsync(12)
 
-      const andThroughResultAsyncFn = jest.fn(() => errAsync('oh no!'))
+      const andThroughResultAsyncFn = vitest.fn(() => errAsync('oh no!'))
 
       const thrued = asyncVal.andThrough(andThroughResultAsyncFn)
 
@@ -995,7 +1018,7 @@ describe('ResultAsync', () => {
     it('Returns the original value when map function returning Result succeeds', async () => {
       const asyncVal = okAsync(12)
 
-      const andThroughResultFn = jest.fn(() => ok('good'))
+      const andThroughResultFn = vitest.fn(() => ok('good'))
 
       const thrued = asyncVal.andThrough(andThroughResultFn)
 
@@ -1011,7 +1034,7 @@ describe('ResultAsync', () => {
     it('Maps to an error when map function returning Result fails', async () => {
       const asyncVal = okAsync(12)
 
-      const andThroughResultFn = jest.fn(() => err('oh no!'))
+      const andThroughResultFn = vitest.fn(() => err('oh no!'))
 
       const thrued = asyncVal.andThrough(andThroughResultFn)
 
@@ -1027,7 +1050,7 @@ describe('ResultAsync', () => {
     it('Skips an Error', async () => {
       const asyncVal = errAsync<string, string>('Wrong format')
 
-      const andThroughResultFn = jest.fn(() => ok<string, string>('good'))
+      const andThroughResultFn = vitest.fn(() => ok<string, string>('good'))
 
       const notMapped = asyncVal.andThrough(andThroughResultFn)
 
@@ -1044,7 +1067,7 @@ describe('ResultAsync', () => {
   describe('andTee', () => {
     it('Calls the passed function but returns an original ok', async () => {
       const okVal = okAsync(12)
-      const passedFn = jest.fn((_number) => {})
+      const passedFn = vitest.fn((_number) => {})
 
       const teed = await okVal.andTee(passedFn)
 
@@ -1054,7 +1077,9 @@ describe('ResultAsync', () => {
     })
     it('returns an original ok even when the passed function fails', async () => {
       const okVal = okAsync(12)
-      const passedFn = jest.fn((_number) => { throw new Error('OMG!') })
+      const passedFn = vitest.fn((_number) => {
+        throw new Error('OMG!')
+      })
 
       const teed = await okVal.andTee(passedFn)
 
@@ -1064,11 +1089,35 @@ describe('ResultAsync', () => {
     })
   })
 
+  describe('orTee', () => {
+    it('Calls the passed function but returns an original err', async () => {
+      const errVal = errAsync(12)
+      const passedFn = vitest.fn((_number) => {})
+
+      const teed = await errVal.orTee(passedFn)
+
+      expect(teed.isErr()).toBe(true)
+      expect(passedFn).toHaveBeenCalledTimes(1)
+      expect(teed._unsafeUnwrapErr()).toStrictEqual(12)
+    })
+    it('returns an original err even when the passed function fails', async () => {
+      const errVal = errAsync(12)
+      const passedFn = vitest.fn((_number) => {
+        throw new Error('OMG!')
+      })
+
+      const teed = await errVal.orTee(passedFn)
+
+      expect(teed.isErr()).toBe(true)
+      expect(passedFn).toHaveBeenCalledTimes(1)
+      expect(teed._unsafeUnwrapErr()).toStrictEqual(12)
+    })
+  })
+
   describe('orElse', () => {
     it('Skips orElse on an Ok value', async () => {
       const okVal = okAsync(12)
-      const errorCallback = jest.fn((_errVal) => errAsync<number, string>('It is now a string'))
-
+      const errorCallback = vitest.fn((_errVal) => errAsync<number, string>('It is now a string'))
 
       const result = await okVal.orElse(errorCallback)
 
@@ -1079,7 +1128,7 @@ describe('ResultAsync', () => {
 
     it('Invokes the orElse callback on an Err value', async () => {
       const myResult = errAsync('BOOOM!')
-      const errorCallback = jest.fn((_errVal) => errAsync(true))
+      const errorCallback = vitest.fn((_errVal) => errAsync(true))
 
       const result = await myResult.orElse(errorCallback)
 
@@ -1089,7 +1138,7 @@ describe('ResultAsync', () => {
 
     it('Accepts a regular Result in the callback', async () => {
       const myResult = errAsync('BOOOM!')
-      const errorCallback = jest.fn((_errVal) => err(true))
+      const errorCallback = vitest.fn((_errVal) => err(true))
 
       const result = await myResult.orElse(errorCallback)
 
@@ -1100,8 +1149,8 @@ describe('ResultAsync', () => {
 
   describe('match', () => {
     it('Matches on an Ok', async () => {
-      const okMapper = jest.fn((_val) => 'weeeeee')
-      const errMapper = jest.fn((_val) => 'wooooo')
+      const okMapper = vitest.fn((_val) => 'weeeeee')
+      const errMapper = vitest.fn((_val) => 'wooooo')
 
       const matched = await okAsync(12).match(okMapper, errMapper)
 
@@ -1111,8 +1160,8 @@ describe('ResultAsync', () => {
     })
 
     it('Matches on an Error', async () => {
-      const okMapper = jest.fn((_val) => 'weeeeee')
-      const errMapper = jest.fn((_val) => 'wooooo')
+      const okMapper = vitest.fn((_val) => 'weeeeee')
+      const errMapper = vitest.fn((_val) => 'wooooo')
 
       const matched = await errAsync('bad').match(okMapper, errMapper)
 
@@ -1196,7 +1245,7 @@ describe('ResultAsync', () => {
 
         return 12
       })
-      
+
       const val = await example()
       expect(val.isErr()).toBe(true)
 
@@ -1206,9 +1255,9 @@ describe('ResultAsync', () => {
     it('Accepts an error handler as a second argument', async () => {
       const example = ResultAsync.fromThrowable(
         () => Promise.reject('No!'),
-        (e) => new Error('Oops: ' + e)
+        (e) => new Error('Oops: ' + e),
       )
-      
+
       const val = await example()
       expect(val.isErr()).toBe(true)